client/requirements.txt
argparse==1.2.1
requests==2.2.1
websocket-client==0.12.0
wsgiref==0.1.2
